whale rssi function

+++ modules/whale_radio.h
@@ -36,6 +36,6 @@ int w_radio_broadcast(
 	ptrdiff_t buffer_size
 );
 
-int w_radio_get_rssi(int16_t *rssi);
+int w_radio_get_rssi(int *rssi);
 
 #endif // WHALE_RADIO_INTERFACE_H
